Moved DTNHostTest to src/test from test/.
